[MAINTENANCE] Move ParallelBigRNN into nlp.model namespace (#1118)
* move parallelbigrnn into api

* Add comments

* fix lint error

* remove parallelbigrnn from the list

* fix pylint errors
